requirements.txt
customtkinter
tkinterdnd2
Pillow
numpy